WIP: First attempt at adding ListableEntityListBase::AddListener
Currently the callback returns a json the contains info about lon/lat of the click (as per [1], click event).
So far I was not able to figure out how to get the GUID of an item using the event.
The event's parameter is a MouseEvent isntance, which gets a GUID, but it seems its the click's GUID rather than the clicked-item's GUID.

[1](https://developers.google.com/maps/documentation/javascript/reference#Circle)